requirements.txt
diffusers
Pillow<10.0
accelerate>=0.11.0
torch>=1.4
torchvision
transformers>=4.21.0
clip-interrogator==0.5.4